Add support for bold, italic, and lists
Added the ability to specify unordered lists, ordered lists, bold and
italic.

All elements (with the exception of bold and itaic) added can be used
anywhere as stated in the specification.  So I've got bold and italic in
paragraphs, and list items.

The only exception at the moment is that bold/italic cannot be nested in
one another.

Changed sample to test out new capabilities.

Changes to be committed:
      modified:   Translate.java
      modified:   sample.xml